+++ NEWS.md
@@ -38,6 +38,8 @@
 
 ## Features and bugfixes
 
+* `arg_exclusive()` checks that only one argument of a set is supplied (#1261).
+
 * `call_name()` and `call_ns()` now fail with calls of the form
   `foo::bar` (#670).